# backend/docs/swagger.yaml
definitions:
...
  domain.CreateMediaRequest:
    properties:
      alt:
        type: string
      filename:
        type: string
      filesize:
        type: integer
      height:
        type: integer
      mime_type:
        type: string
      url:
        type: string
      width:
        type: integer
    required:
    - alt
    - filename
    - url
    type: object
...
  domain.Media:
    properties:
      alt:
        type: string
      created_at:
        type: string
      filename:
        type: string
      filesize:
        type: integer
      height:
        type: integer
      id:
        type: integer
      mime_type:
        type: string
      updated_at:
        type: string
      url:
        type: string
      width:
        type: integer
    type: object
...
  domain.UpdateMediaRequest:
    properties:
      alt:
        type: string
      filename:
        type: string
      filesize:
        type: integer
      height:
        type: integer
      mime_type:
        type: string
      url:
        type: string
      width:
        type: integer
    required:
    - alt
    - filename
    - url
    type: object
...
  response.ListMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  response.ListResponse-domain_Media:
    properties:
      code:
        type: string
      data:
        items:
          $ref: '#/definitions/domain.Media'
        type: array
      message:
        type: string
      meta:
        $ref: '#/definitions/response.ListMeta'
      success:
        type: boolean
    type: object
...
  response.Response-any:
    properties:
      code:
        type: string
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  response.Response-domain_Media:
    properties:
      code:
        type: string
      data:
        $ref: '#/definitions/domain.Media'
      message:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact: {}
  description: Customer auth, product catalog, and admin CMS API for the Diagonals
    website backend.
  title: Diagonals API
  version: "1.0"
paths:
...
  /api/v1/media:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all media files
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ListResponse-domain_Media'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: Get all media
      tags:
      - Media
    post:
      consumes:
      - application/json
      description: Create a new media file record
      parameters:
      - description: Media payload
        in: body
        name: media
        required: true
        schema:
          $ref: '#/definitions/domain.CreateMediaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response-domain_Media'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: Create media
      tags:
      - Media
  /api/v1/media/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a media file record by ID
      parameters:
      - description: Media ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: Delete media
      tags:
      - Media
    get:
      consumes:
      - application/json
      description: Retrieve a single media file by ID
      parameters:
      - description: Media ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response-domain_Media'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: Get media by ID
      tags:
      - Media
    put:
      consumes:
      - application/json
      description: Update an existing media file record
      parameters:
      - description: Media ID
        in: path
        name: id
        required: true
        type: integer
      - description: Media payload
        in: body
        name: media
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateMediaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response-domain_Media'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: Update media
      tags:
      - Media
  /api/v1/media/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload a media file via multipart/form-data, save it on the backend
        file store, and create a media record
      parameters:
      - description: Media file
        in: formData
        name: file
        required: true
        type: file
      - description: Alt text
        in: formData
        name: alt
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response-domain_Media'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response-any'
      summary: Upload media file
      tags:
      - Media
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"